[FEAT] isLiked 필드 추가
