+++ src/main/java/com/growplan/user/domain/User.java
@@ -3,7 +3,6 @@ package com.growplan.user.domain;
 import com.growplan.child.domain.UserChild;
 import com.growplan.common.BaseEntity;
 import com.growplan.common.type.StatusType;
-import com.growplan.like.domain.RecordLike;
 import com.growplan.login.domain.UserSign;
 import com.growplan.login.domain.type.SocialLoginType;
 import jakarta.persistence.*;
@@ -60,9 +59,6 @@ public class User extends BaseEntity {
 
     @OneToMany(mappedBy = "user", orphanRemoval = true, cascade = CascadeType.ALL)
     private List<UserChild> userChildren = new ArrayList<>();
-
-    @OneToMany(mappedBy = "user", orphanRemoval = true, cascade = CascadeType.ALL)
-    private List<RecordLike> recordLikes = new ArrayList<>();
 
     @OneToOne(orphanRemoval = true, cascade = CascadeType.ALL)
     private UserSign userSign;
